requirements.txt
setuptools>=44.0.0
numpy~=1.22.1
PyQt5~=5.15.6
pyqtgraph~=0.12.3
pillow~=9.0.1